+++ packages/pros-async/src/lib.rs
@@ -9,14 +9,16 @@
 
 extern crate alloc;
 
-use core::{future::Future, task::Poll};
+use core::future::Future;
 
 use async_task::Task;
-use executor::EXECUTOR;
 use pros_core::error::Result;
 
 mod executor;
+pub mod futures;
 mod reactor;
+
+pub use futures::*;
 
 /// Runs a future in the background without having to await it
 /// To get the the return value you can await a task.
@@ -28,40 +30,6 @@ pub fn spawn<T>(future: impl Future<Output = T> + 'static) -> Task<T> {
 /// Does not poll all futures to completion.
 pub fn block_on<F: Future + 'static>(future: F) -> F::Output {
     executor::EXECUTOR.with(|e| e.block_on(spawn(future)))
-}
-
-/// A future that will complete after the given duration.
-/// Sleep futures that are closer to completion are prioritized to improve accuracy.
-#[derive(Debug)]
-pub struct SleepFuture {
-    target_millis: u32,
-}
-impl Future for SleepFuture {
-    type Output = ();
-
-    fn poll(
-        self: core::pin::Pin<&mut Self>,
-        cx: &mut core::task::Context<'_>,
-    ) -> core::task::Poll<Self::Output> {
-        if self.target_millis < unsafe { pros_sys::millis() } {
-            Poll::Ready(())
-        } else {
-            EXECUTOR.with(|e| {
-                e.reactor
-                    .borrow_mut()
-                    .sleepers
-                    .push(cx.waker().clone(), self.target_millis)
-            });
-            Poll::Pending
-        }
-    }
-}
-
-/// Returns a future that will complete after the given duration.
-pub fn sleep(duration: core::time::Duration) -> SleepFuture {
-    SleepFuture {
-        target_millis: unsafe { pros_sys::millis() + duration.as_millis() as u32 },
-    }
 }
 
 /// A trait for robot code that spins up the pros-rs async executor.